fix mapper process arror

+++ src/main/java/com/example/mapstruct/models/Employee.java
@@ -4,7 +4,6 @@ package com.example.mapstruct.models;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -19,7 +18,7 @@ public class Employee {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private int id;
+    private Long id;
     private String firstName;
     private String lastName;
     private String jobRole;
